+++ CHANGES.rst
@@ -4,6 +4,9 @@ unreleased
 Features
 --------
 
+- Add `Configurator.resolve_asset` and `Request.resolve_asset` methods.  See
+  https://github.com/Pylons/pyramid/pull/3745.
+
 Bug Fixes
 ---------